Added some cleanup
