+++ knora/models/resourceclass.py
@@ -18,6 +18,12 @@ class SetEncoder(json.JSONEncoder):
             return list(obj)
         return json.JSONEncoder.default(self, obj)
 
+"""
+This model implements the handling of resource classes. It contains two classes that work closely together:
+    * "HasProperty" deals with the association of Property-instances with the Resource-instances. This association
+      is done using the "cardinality"-clause
+    *  "ResourceClass" is the main class representing a knora resource class.
+"""
 @strict
 class HasProperty:
     class Ptype(Enum):
@@ -291,17 +297,96 @@ class HasProperty:
 
 @strict
 class ResourceClass:
+    """
+    This class represents a knora resource class
+
+    Attributes
+    ----------
+
+    con : Connection
+        A Connection instance to a Knora server
+
+    id : str
+        IRI of the ResourceClass [readonly, cannot be modified after creation of instance]
+
+    name: str
+        The name of the resource class, e.g. "Book", "Person", "Portait". Usually these names start
+        with a capital letter
+
+    ontology_id: str
+        The IRI/Id of the ontology this resource class belongs to
+
+    superlcasses: str, List[str]
+        This is a list of superclasses for this resource class. Usually a project specific class must at least
+        be a subclass of "Resource", but can be subclassed of any other valid resource class. In addition, external
+        ontologies may be referenced:
+        e.g.:
+        ```
+        "super": "Resource"
+
+        "super": ":MySpecialTYpe"
+
+        "super": [":MySpecialType", "dcterms:BibliographicResource"]
+        ```
+
+    label: language dependent string, that is a dict like {"en": "Biblography", "de": "Literaturverzeichnis"}
+        A label (human readable name) for the resource
+
+    comment: language dependent string, that is a dict like {"en": "a comment", "de": "Ein Kommentar"}
+        A commentary to further explain what this resource class is used for
+
+    permission: str
+        The default permissions to be used if an instance of this resource class is being created
+
+    has_properties: Dict[str, HasProperty]
+        Holds a dict with the property names as keys and a HasProperty instance. HasProperty holds
+        the information, how this resource class uses this property (basically the cardinality)
+
+    changed: bool
+        is set to True, if one of the fields has been chaned by the user (internal use only!)
+
+
+    Methods
+    -------
+
+    addLabel: Add a new label to the resource
+        addLabel(self, lang: Union[Languages, str], value: str) -> None
+
+    getProperty: Get information about a property defined for this resource class
+        getPropertery(prop_id: str) -> HasProperty
+        returns a HasProperty-instance
+
+    addProperty: Add a new property to the resource class
+        addProperty(property_id: str, cardinality: Cardinality, last_modification_date: LastModificationDate)
+        -> Optional[LastModificationDate]
+
+    updateProperty: Updates the cardinality parameters of the given property with the resource class
+        updateProperty(self, property_id: str, cardinality: Cardinality, last_modification_date: LastModificationDate)
+        -> Optional[LastModificationDate]
+        Please note that the cardinality usually can only be changed to be *less* restrictive!
+
+    create: Create a new resource class on the connected server
+
+    update: Update the information of a resource class on the connected server
+
+    delete: Mark a resource class as deleted (on the connected server)
+
+    createDefinitionFileObj: Create an object suitable for jsonification that conforms the the knora-py ontology tools
+
+    print: Print the content of this class to the console
+
+    """
     con: Connection
-    _context: Context
-    _id: str
-    _name: str
-    _ontology_id: str
-    _superclasses: List[str]
-    _label: LangString
-    _comment: LangString
-    _permissions: str
-    _has_properties: Dict[str, HasProperty]
-    _changed: Set[str]
+    __context: Context
+    __id: str
+    __name: str
+    __ontology_id: str
+    __superclasses: List[str]
+    __label: LangString
+    __comment: LangString
+    __permissions: str
+    ___has_properties: Dict[str, HasProperty]
+    __changed: Set[str]
 
     def __init__(self,
                  con: Connection,
@@ -314,53 +399,67 @@ class ResourceClass:
                  comment: Optional[Union[LangString, str]] = None,
                  permissions: Optional[str] = None,
                  has_properties: Optional[Dict[str, HasProperty]] = None):
+        """
+        Create an instance of  ResourceClass
+
+        :param con:
+        :param context:
+        :param id:
+        :param name:
+        :param ontology_id:
+        :param superclasses:
+        :param label:
+        :param comment:
+        :param permissions:
+        :param has_properties:
+        """
         if not isinstance(con, Connection):
             raise BaseError('"con"-parameter must be an instance of Connection')
         if not isinstance(context, Context):
             raise BaseError('"context"-parameter must be an instance of Context')
         self.con = con
-        self._context = context
-        self._id = id
-        self._name = name
-        self._ontology_id = ontology_id
+        self.__context = context
+        self.__id = id
+        self.__name = name
+        self.__ontology_id = ontology_id
         if isinstance(superclasses, ResourceClass):
-            self._superclasses = list(map(lambda a: a.id, superclasses))
+            self.__superclasses = list(map(lambda a: a.id, superclasses))
         else:
-            self._superclasses = superclasses
+            self.__superclasses = superclasses
         #
         # process label
         #
         if label is not None:
             if isinstance(label, str):
-                self._label = LangString(label)
+                self.__label = LangString(label)
             elif isinstance(label, LangString):
-                self._label = label
+                self.__label = label
             else:
                 raise BaseError('Invalid LangString for label!')
         else:
-            self._label = None
+            self.__label = None
         #
         # process comment
         #
         if comment is not None:
             if isinstance(comment, str):
-                self._comment = LangString(comment)
+                self.__comment = LangString(comment)
             elif isinstance(comment, LangString):
-                self._comment = comment
+                self.__comment = comment
             else:
                 raise BaseError('Invalid LangString for comment!')
         else:
-            self._comment = None
-        self._permissions = permissions
-        self._has_properties = has_properties
-        self._changed = set()
+            self.__comment = None
+        self.__permissions = permissions
+        self.___has_properties = has_properties
+        self.__changed = set()
 
     #
     # Here follows a list of getters/setters
     #
     @property
     def name(self) -> Optional[str]:
-        return self._name
+        return self.__name
 
     @name.setter
     def name(self, value: str) -> None:
@@ -368,7 +467,7 @@ class ResourceClass:
 
     @property
     def id(self) -> Optional[str]:
-        return self._id
+        return self.__id
 
     @id.setter
     def id(self, value: str) -> None:
@@ -376,7 +475,7 @@ class ResourceClass:
 
     @property
     def ontology_id(self) -> Optional[str]:
-        return self._ontology_id
+        return self.__ontology_id
 
     @ontology_id.setter
     def ontology_id(self, value: str) -> None:
@@ -384,7 +483,7 @@ class ResourceClass:
 
     @property
     def superclasses(self) -> Optional[List[str]]:
-        return self._superclasses
+        return self.__superclasses
 
     @superclasses.setter
     def superclasses(self, value: List[str]) -> None:
@@ -392,58 +491,58 @@ class ResourceClass:
 
     @property
     def label(self) -> Optional[LangString]:
-        return self._label
+        return self.__label
 
     @label.setter
     def label(self, value: Optional[Union[LangString, str]]) -> None:
         if value is None:
-            self._label.empty()  # clear all labels
+            self.__label.empty()  # clear all labels
         else:
             if isinstance(value, LangString):
-                self._label = value
+                self.__label = value
             elif isinstance(value, str):
-                self._label = LangString(value)
+                self.__label = LangString(value)
             else:
                 raise BaseError('Not a valid LangString')
-        self._changed.add('label')
+        self.__changed.add('label')
 
 
     def addLabel(self, lang: Union[Languages, str], value: str) -> None:
-        self._label[lang] = value
-        self._changed.add('label')
+        self.__label[lang] = value
+        self.__changed.add('label')
 
     def rmLabel(self, lang: Union[Languages, str]) -> None:
-        del self._label[lang]
-        self._changed.add('label')
+        del self.__label[lang]
+        self.__changed.add('label')
 
     @property
     def comment(self) -> Optional[LangString]:
-        return self._comment
+        return self.__comment
 
     @comment.setter
     def comment(self, value: Optional[LangString]) -> None:
         if value is None:
-            self._comment.empty()  # clear all comments!
+            self.__comment.empty()  # clear all comments!
         else:
             if isinstance(value, LangString):
-                self._comment = value
+                self.__comment = value
             elif isinstance(value, str):
-                self._comment = LangString(value)
+                self.__comment = LangString(value)
             else:
                 raise BaseError('Not a valid LangString')
-        self._changed.add('comment')
+        self.__changed.add('comment')
 
     def addComment(self, lang: Union[Languages, str], value: str) -> None:
-        self._comment[lang] = value
-        self._changed.add('comment')
+        self.__comment[lang] = value
+        self.__changed.add('comment')
 
     def rmComment(self, lang: Union[Languages, str]) -> None:
-        del self._comment[lang]
-        self._changed.add('comment')
+        del self.__comment[lang]
+        self.__changed.add('comment')
 
     @property
     def permissions(self) -> Optional[str]:
-        return self._permissions
+        return self.__permissions
 
     @permissions.setter
     def permissions(self, value: str) -> None:
@@ -451,46 +550,46 @@ class ResourceClass:
 
     @property
     def has_property(self) -> Dict[str, HasProperty]:
-        return self._has_properties
+        return self.___has_properties
 
     @has_property.setter
     def has_property(self, value: str) -> None:
         raise BaseError('"has_property" cannot be modified!')
 
     def getProperty(self, property_id) -> Optional[HasProperty]:
-        if self._has_properties is None:
+        if self.___has_properties is None:
             return None
         else:
-            return self._has_properties.get(self._context.get_prefixed_iri(property_id))
+            return self.___has_properties.get(self.__context.get_prefixed_iri(property_id))
 
     def addProperty(self, property_id: str, cardinality: Cardinality, last_modification_date: LastModificationDate) -> Optional[LastModificationDate]:
-        if self._has_properties.get(property_id) is None:
+        if self.___has_properties.get(property_id) is None:
             latest_modification_date, resclass = HasProperty(con=self.con,
-                                                             context=self._context,
-                                                             ontology_id=self._ontology_id,
+                                                             context=self.__context,
+                                                             ontology_id=self.__ontology_id,
                                                              property_id=property_id,
                                                              resclass_id=self.id,
                                                              cardinality=cardinality).create(last_modification_date)
             hp = resclass.getProperty(property_id)
-            hp.ontology_id = self._context.iri_from_prefix(self._ontology_id)
+            hp.ontology_id = self.__context.iri_from_prefix(self.__ontology_id)
             hp.resclass_id = self.id
-            self._has_properties[hp.property_id] = hp
+            self.___has_properties[hp.property_id] = hp
             return latest_modification_date
         else:
             raise BaseError("Property already has cardinality in this class!")
 
     def updateProperty(self, property_id: str, cardinality: Cardinality, last_modification_date: LastModificationDate) -> Optional[LastModificationDate]:
-        property_id = self._context.get_prefixed_iri(property_id)
-        if self._has_properties.get(property_id) is not None:
-            has_properties = self._has_properties[property_id]
+        property_id = self.__context.get_prefixed_iri(property_id)
+        if self.___has_properties.get(property_id) is not None:
+            has_properties = self.___has_properties[property_id]
             onto_id = has_properties.ontology_id  # save for later user
             rescl_id = has_properties.resclass_id  # save for later user
             has_properties.cardinality = cardinality
             latest_modification_date, resclass = has_properties.update(last_modification_date)
             hp = resclass.getProperty(property_id)
-            hp.ontology_id = self._context.iri_from_prefix(onto_id)  # restore value
+            hp.ontology_id = self.__context.iri_from_prefix(onto_id)  # restore value
             hp.resclass_id = rescl_id  # restore value
-            self._has_properties[hp.property_id] = hp
+            self.___has_properties[hp.property_id] = hp
             return latest_modification_date
         else:
             return last_modification_date
@@ -549,29 +648,29 @@ class ResourceClass:
                 if tmp[0]:
                     return {"@id": resref}  # fully qualified name in the form "prefix:name"
                 else:
-                    return {"@id": self._context.prefix_from_iri(self._ontology_id) + ':' + tmp[1]}  # ":name" in current ontology
+                    return {"@id": self.__context.prefix_from_iri(self.__ontology_id) + ':' + tmp[1]}  # ":name" in current ontology
             else:
                 return {"@id": "knora-api:" + resref}  # no ":", must be from knora-api!
 
         tmp = {}
         exp = re.compile('^http.*')  # It is already a fully IRI
-        if exp.match(self._ontology_id):
-            resid = self._context.prefix_from_iri(self._ontology_id) + ":" + self._name
-            ontid = self._ontology_id
+        if exp.match(self.__ontology_id):
+            resid = self.__context.prefix_from_iri(self.__ontology_id) + ":" + self.__name
+            ontid = self.__ontology_id
         else:
-            resid = self._ontology_id + ":" + self._name
-            ontid = self._context.iri_from_prefix(self._ontology_id)
+            resid = self.__ontology_id + ":" + self.__name
+            ontid = self.__context.iri_from_prefix(self.__ontology_id)
         if action == Actions.Create:
-            if self._name is None:
+            if self.__name is None:
                 raise BaseError("There must be a valid resource class name!")
-            if self._ontology_id is None:
+            if self.__ontology_id is None:
                 raise BaseError("There must be a valid ontology_id given!")
-            if self._superclasses is None:
+            if self.__superclasses is None:
                 superclasses = [{"@id": "knora-api:Resource"}]
             else:
-                superclasses = list(map(resolve_resref, self._superclasses))
-            if self._comment is None or self._comment.isEmpty():
-                self._comment = LangString("no comment vailable")
+                superclasses = list(map(resolve_resref, self.__superclasses))
+            if self.__comment is None or self.__comment.isEmpty():
+                self.__comment = LangString("no comment vailable")
             tmp = {
                 "@id": ontid,  # self._ontology_id,
                 "@type": "owl:Ontology",
@@ -579,11 +678,11 @@ class ResourceClass:
                 "@graph": [{
                     "@id": resid,
                     "@type": "owl:Class",
-                    "rdfs:label": self._label.toJsonLdObj(),
-                    "rdfs:comment": self._comment.toJsonLdObj(),
+                    "rdfs:label": self.__label.toJsonLdObj(),
+                    "rdfs:comment": self.__comment.toJsonLdObj(),
                     "rdfs:subClassOf": superclasses
                 }],
-                "@context": self._context.toJsonObj(),
+                "@context": self.__context.toJsonObj(),
             }
         elif action == Actions.Update:
             tmp = {
@@ -594,14 +693,14 @@ class ResourceClass:
                     "@id": resid,
                     "@type": "owl:Class",
                 }],
-                "@context": self._context.toJsonObj(),
+                "@context": self.__context.toJsonObj(),
             }
             if what == 'label':
-                if not self._label.isEmpty() and 'label' in self._changed:
-                    tmp['@graph'][0]['rdfs:label'] = self._label.toJsonLdObj()
+                if not self.__label.isEmpty() and 'label' in self.__changed:
+                    tmp['@graph'][0]['rdfs:label'] = self.__label.toJsonLdObj()
             if what == 'comment':
-                if not self._comment.isEmpty() and 'comment' in self._changed:
-                    tmp['@graph'][0]['rdfs:comment'] = self._comment.toJsonLdObj()
+                if not self.__comment.isEmpty() and 'comment' in self.__changed:
+                    tmp['@graph'][0]['rdfs:comment'] = self.__comment.toJsonLdObj()
         return tmp
 
     def create(self, last_modification_date: LastModificationDate) -> Tuple[LastModificationDate, 'ResourceClass']:
@@ -609,7 +708,7 @@ class ResourceClass:
         jsondata = json.dumps(jsonobj)
         result = self.con.post('/v2/ontologies/classes', jsondata)
         last_modification_date = LastModificationDate(result['knora-api:lastModificationDate'])
-        return last_modification_date, ResourceClass.fromJsonObj(self.con, self._context, result['@graph'])
+        return last_modification_date, ResourceClass.fromJsonObj(self.con, self.__context, result['@graph'])
 
     def update(self, last_modification_date: LastModificationDate) -> Tuple[LastModificationDate, 'ResourceClass']:
         #
@@ -617,45 +716,45 @@ class ResourceClass:
         #
         result = None
         something_changed = False
-        if 'label' in self._changed:
+        if 'label' in self.__changed:
             jsonobj = self.toJsonObj(last_modification_date, Actions.Update, 'label')
             jsondata = json.dumps(jsonobj, cls=SetEncoder, indent=4)
             result = self.con.put('v2/ontologies/classes', jsondata)
             last_modification_date = LastModificationDate(result['knora-api:lastModificationDate'])
             something_changed = True
-        if 'comment' in self._changed:
+        if 'comment' in self.__changed:
             jsonobj = self.toJsonObj(last_modification_date, Actions.Update, 'comment')
             jsondata = json.dumps(jsonobj, cls=SetEncoder, indent=4)
             result = self.con.put('v2/ontologies/classes', jsondata)
             last_modification_date = LastModificationDate(result['knora-api:lastModificationDate'])
             something_changed = True
         if something_changed:
-            return last_modification_date, ResourceClass.fromJsonObj(self.con, self._context, result['@graph'])
+            return last_modification_date, ResourceClass.fromJsonObj(self.con, self.__context, result['@graph'])
         else:
             return last_modification_date, self
 
     def delete(self, last_modification_date: LastModificationDate) -> LastModificationDate:
-        result = self.con.delete('v2/ontologies/classes/' + quote_plus(self._id) + '?lastModificationDate=' + str(last_modification_date))
+        result = self.con.delete('v2/ontologies/classes/' + quote_plus(self.__id) + '?lastModificationDate=' + str(last_modification_date))
         return LastModificationDate(result['knora-api:lastModificationDate'])
 
     def createDefinitionFileObj(self, context: Context, shortname: str, skiplist: List[str]):
         resource = {
-            "name": self._name,
-            "labels": self._label.createDefinitionFileObj(),
+            "name": self.__name,
+            "labels": self.__label.createDefinitionFileObj(),
         }
-        if self._comment:
-            resource["comments"] = self._comment.createDefinitionFileObj()
-        if self._superclasses:
-            if len(self._superclasses) > 1:
+        if self.__comment:
+            resource["comments"] = self.__comment.createDefinitionFileObj()
+        if self.__superclasses:
+            if len(self.__superclasses) > 1:
                 superclasses = []
-                for sc in self._superclasses:
+                for sc in self.__superclasses:
                     superclasses.append(context.reduce_iri(sc, shortname))
             else:
-                superclasses = context.reduce_iri(self._superclasses[0], shortname)
+                superclasses = context.reduce_iri(self.__superclasses[0], shortname)
             resource["super"] = superclasses
-        if self._has_properties:
+        if self.___has_properties:
             cardinalities = []
-            for pid, hp in self._has_properties.items():
+            for pid, hp in self.___has_properties.items():
                 if hp.property_id in skiplist:
                     print("Skip ", hp.property_id)
                     continue
@@ -668,22 +767,22 @@ class ResourceClass:
     def print(self, offset: int = 0):
         blank = ' '
         print(f'{blank:>{offset}}Resource Class Info')
-        print(f'{blank:>{offset+2}}Name:            {self._name}')
-        print(f'{blank:>{offset+2}}Ontology prefix: {self._ontology_id}')
+        print(f'{blank:>{offset+2}}Name:            {self.__name}')
+        print(f'{blank:>{offset+2}}Ontology prefix: {self.__ontology_id}')
         print(f'{blank:>{offset+2}}Superclasses:')
-        if self._superclasses is not None:
-            for sc in self._superclasses:
+        if self.__superclasses is not None:
+            for sc in self.__superclasses:
                 print(f'{blank:>{offset + 4}}{sc}')
-        if self._label is not None:
+        if self.__label is not None:
             print(f'{blank:>{offset + 2}}Labels:')
-            self._label.print(offset + 4)
-        if self._comment is not None:
+            self.__label.print(offset + 4)
+        if self.__comment is not None:
             print(f'{blank:>{offset + 2}}Comments:')
-            self._comment.print(offset + 4)
-        if self._has_properties is not None:
+            self.__comment.print(offset + 4)
+        if self.___has_properties is not None:
             print(f'{blank:>{offset + 2}}Has properties:')
-            if self._has_properties is not None:
-                for pid, hp in self._has_properties.items():
+            if self.___has_properties is not None:
+                for pid, hp in self.___has_properties.items():
                     hp.print(offset + 4)
