+++ knora/onto_commons.py
@@ -1,0 +1,38 @@
+import os
+from typing import List, Set, Dict, Tuple, Optional
+from pprint import pprint
+import argparse
+import json
+from jsonschema import validate
+import sys
+import re
+
+from models.helpers import Actions, BaseError, Context, Cardinality
+from models.langstring import Languages, LangStringParam, LangString
+from models.connection import Connection, Error
+from models.project import Project
+from models.listnode import ListNode
+from models.group import Group
+from models.user import User
+from models.ontology import Ontology
+from models.propertyclass import PropertyClass
+from models.resourceclass import ResourceClass
+
+def list_creator(con: Connection, project: Project, parent_node: ListNode, nodes: List[dict]):
+    nodelist = []
+    for node in nodes:
+        newnode = ListNode(
+            con=con,
+            project=project,
+            label=node["labels"],
+            comment=node.get("comments"),
+            name=node["name"],
+            parent=parent_node
+        ).create()
+        if node.get('nodes') is not None:
+            subnodelist = list_creator(con, project, newnode, node['nodes'])
+            nodelist.append({newnode.name: {"id": newnode.id, 'nodes': subnodelist}})
+        else:
+            nodelist.append({newnode.name: {"id": newnode.id}})
+    return nodelist
+
